dmt api change

+++ org.osgi.impl.service.policy/src/org/osgi/impl/service/policy/condpermadmin/ConditionInfoMetaNode.java
@@ -36,8 +36,8 @@ public final class ConditionInfoMetaNode implements DmtMetaNode {
 	public int getMaxOccurrence() {	return 1; }
 	public boolean isZeroOccurrenceAllowed() { return false; }
 	public DmtData getDefault() { return null; }
-	public int getMax() { return Integer.MAX_VALUE; }
-	public int getMin() { return Integer.MIN_VALUE;	}
+	public double getMax() { return Double.MAX_VALUE; }
+	public double getMin() { return Double.MIN_VALUE;	}
 	public DmtData[] getValidValues() { return null; }
 	public int getFormat() { return DmtData.FORMAT_STRING; }
 	public String[] getMimeTypes() { return null; }
