+++ org.osgi.impl.service.policy/src/org/osgi/impl/service/policy/permadmin/DefaultMetaNode.java
@@ -44,8 +44,8 @@ public final class DefaultMetaNode implements DmtMetaNode {
 	public int getMaxOccurrence() {	return 1; }
 	public boolean isZeroOccurrenceAllowed() { return true; }
 	public DmtData getDefault() { return defaultData; }
-	public int getMax() { return Integer.MAX_VALUE; }
-	public int getMin() { return Integer.MIN_VALUE;	}
+	public double getMax() { return Double.MAX_VALUE; }
+	public double getMin() { return Double.MIN_VALUE;	}
 	public DmtData[] getValidValues() { return null; }
 	public int getFormat() { return DmtData.FORMAT_STRING; }
 	public String[] getMimeTypes() { return null; }
